Use double rather than qreal in remaining tools code.
This changes the metadata fields that were qreal to double, which
results in knock-on changes in ld-analyse, ld-process-vits and
ld-dropout-correct. Following ld-decode's usual style of using
static_cast even when an implicit conversion would do, I've added a few
missing casts and removed about an equal number of unnecessary ones.

The only remaining use of qreal is in ld-analyse, where it uses QPointF.

+++ tools/ld-analyse/dropoutanalysisdialog.cpp
@@ -86,9 +86,9 @@ void DropoutAnalysisDialog::removeChartContents()
 }
 
 // Add a data point to the chart
-void DropoutAnalysisDialog::addDataPoint(qint32 frameNumber, qreal doLength)
+void DropoutAnalysisDialog::addDataPoint(qint32 frameNumber, double doLength)
 {
-    points->append(QPointF(frameNumber, doLength));
+    points->append(QPointF(static_cast<qreal>(frameNumber), static_cast<qreal>(doLength)));
 
     // Keep track of the maximum Y value
     if (doLength > maxY) maxY = doLength;
