+++ tools/ld-analyse/tbcsource.cpp
@@ -301,25 +301,25 @@ qint32 TbcSource::getFrameWidth()
 }
 
 // Get black SNR data for graphing
-QVector<qreal> TbcSource::getBlackSnrGraphData()
+QVector<double> TbcSource::getBlackSnrGraphData()
 {
     return blackSnrGraphData;
 }
 
 // Get white SNR data for graphing
-QVector<qreal> TbcSource::getWhiteSnrGraphData()
+QVector<double> TbcSource::getWhiteSnrGraphData()
 {
     return whiteSnrGraphData;
 }
 
 // Get dropout data for graphing
-QVector<qreal> TbcSource::getDropOutGraphData()
+QVector<double> TbcSource::getDropOutGraphData()
 {
     return dropoutGraphData;
 }
 
 // Get visible dropout data for graphing
-QVector<qreal> TbcSource::getVisibleDropOutGraphData()
+QVector<double> TbcSource::getVisibleDropOutGraphData()
 {
     return visibleDropoutGraphData;
 }
@@ -763,15 +763,15 @@ void TbcSource::generateData()
 
     const qint32 numFrames = ldDecodeMetaData.getNumberOfFrames();
     for (qint32 frameNumber = 0; frameNumber < numFrames; frameNumber++) {
-        qreal doLength = 0;
-        qreal visibleDoLength = 0;
-        qreal blackSnrTotal = 0;
-        qreal whiteSnrTotal = 0;
+        double doLength = 0;
+        double visibleDoLength = 0;
+        double blackSnrTotal = 0;
+        double whiteSnrTotal = 0;
 
         // SNR data may be missing in some fields, so we count the points to prevent
         // the frame average from being thrown-off by missing data
-        qreal blackSnrPoints = 0;
-        qreal whiteSnrPoints = 0;
+        double blackSnrPoints = 0;
+        double whiteSnrPoints = 0;
 
         const LdDecodeMetaData::Field &firstField = ldDecodeMetaData.getField(ldDecodeMetaData.getFirstFieldNumber(frameNumber + 1));
         const LdDecodeMetaData::Field &secondField = ldDecodeMetaData.getField(ldDecodeMetaData.getSecondFieldNumber(frameNumber + 1));
@@ -780,7 +780,7 @@ void TbcSource::generateData()
         if (firstField.dropOuts.size() > 0) {
             // Calculate the total length of the dropouts
             for (qint32 i = 0; i < firstField.dropOuts.size(); i++) {
-                doLength += firstField.dropOuts.endx(i) - firstField.dropOuts.startx(i);
+                doLength += static_cast<double>(firstField.dropOuts.endx(i) - firstField.dropOuts.startx(i));
             }
         }
 
@@ -788,7 +788,7 @@ void TbcSource::generateData()
         if (secondField.dropOuts.size() > 0) {
             // Calculate the total length of the dropouts
             for (qint32 i = 0; i < secondField.dropOuts.size(); i++) {
-                doLength += secondField.dropOuts.endx(i) - secondField.dropOuts.startx(i);
+                doLength += static_cast<double>(secondField.dropOuts.endx(i) - secondField.dropOuts.startx(i));
             }
         }
 
@@ -807,7 +807,7 @@ void TbcSource::generateData()
                         if (firstField.dropOuts.endx(i) < videoParameters.activeVideoEnd) endx = firstField.dropOuts.endx(i);
                         else endx = videoParameters.activeVideoEnd;
 
-                        visibleDoLength += endx - startx;
+                        visibleDoLength += static_cast<double>(endx - startx);
                     }
                 }
             }
@@ -826,7 +826,7 @@ void TbcSource::generateData()
                         if (secondField.dropOuts.endx(i) < videoParameters.activeVideoEnd) endx = secondField.dropOuts.endx(i);
                         else endx = videoParameters.activeVideoEnd;
 
-                        visibleDoLength += endx - startx;
+                        visibleDoLength += static_cast<double>(endx - startx);
                     }
                 }
             }
